+++ allauth/idp/oidc/internal/clientkit.py
@@ -5,6 +5,7 @@ from re import Pattern
 from urllib.parse import ParseResult, parse_qsl, urlparse
 
 from django.core.exceptions import ValidationError
+from django.forms import URLField
 from django.utils.translation import gettext_lazy as _
 
 from allauth.idp.oidc import app_settings
@@ -153,6 +154,8 @@ def clean_post_logout_redirect_uri(
     parameter or via another mechanism. An id_token_hint is also RECOMMENDED
     when this parameter is included.
     """
+    if not post_logout_redirect_uri:
+        return None
     allowed_schemes = {"https"}
     if client:
         allowed_schemes.update(get_used_schemes(client))
@@ -160,7 +163,9 @@ def clean_post_logout_redirect_uri(
             allowed_schemes.add("http")
     parsed = urlparse(post_logout_redirect_uri)
     if not parsed.scheme or parsed.scheme not in allowed_schemes:
-        return None
+        raise ValidationError(URLField.default_error_messages["invalid"])
+    if parsed.scheme in ("http", "https"):
+        URLField().clean(post_logout_redirect_uri)
     return post_logout_redirect_uri